teams_chat/requirements.txt
botbuilder-core>=4.16.0,<5
botbuilder-integration-aiohttp>=4.16.0,<5
aiohttp>=3.9,<4
python-dotenv>=1.0
httpx>=0.27,<1
